requirements.txt
aiohttp
beautifulsoup4
fake_useragent
gpt4all
langchain
numpy
openai
protobuf
PyAudio
python-dotenv
pyttsx3
requests
tiktoken
vosk
google-generativeai
groq
